requirements.txt
attrs==21.2.0
backcall==0.2.0
click==8.0.3
decorator==5.1.0
environs==9.3.5
flake8==3.9.2
Flask==2.0.2
iniconfig==1.1.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
mccabe==0.6.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pymongo==3.12.1
pyparsing==2.4.7
python-dotenv==0.19.2
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.2
